removing the old stuff in favor of the new
