+++ cli/src/main.ts
@@ -8,8 +8,44 @@ import type { Readable as IsomorphicReadable } from 'readable-stream'
 
 import { DwnHttpClient } from '../../src/dwn-http-client.js'
 import { DwnProxy } from '../../src/dwn-proxy.js'
-import { DwnProxyMarkup } from '../../src/dwn-proxy-markup.js'
 import { readReq } from '../../src/dwn-http-server.js'
+
+export class DwnProxyMarkup {
+  static resolveDotDelimited = (obj: any, value: string) => {
+    const propChain = value.split('.').slice(1)
+    let protoValue = obj
+    for (let prop of propChain) {
+      protoValue = protoValue[prop]
+      if (protoValue === undefined) {
+        return false
+      }
+    }
+    return protoValue
+  }
+
+  static populate = (obj: any, replacements: any): any => {
+    for (const [key, value] of Object.entries(obj)) {
+      if (value) {
+        if (Array.isArray(value)) {
+          obj[key] = value.map(item => DwnProxyMarkup.populate(item, replacements))
+        } else if (typeof value === 'object') {
+          obj[key] = DwnProxyMarkup.populate(value, replacements)
+        } else if (typeof value === 'string' && value[0] === '#') {
+          const dotDelimited = value.split('.')
+          if (dotDelimited.length === 1) {
+            if (replacements[value])
+              obj[key] = replacements[value]
+          } else {
+            if (replacements[dotDelimited[0]])
+              obj[key] = DwnProxyMarkup.resolveDotDelimited(replacements[dotDelimited[0]], value)
+          }
+        }
+      }
+    }
+
+    return obj
+  }
+}
 
 type Config = {
   definition: string | any
