--- examples/tbdex-proxy.js
@@ -1,3 +1,0 @@
-import { main } from '../dist/esm/main.mjs'
-
-main()
