+++ src/main.ts
@@ -1,7 +1,6 @@
 export * from './dwn-http-client.js'
 export * from './dwn-http-server.js'
 export * from './dwn-json-rpc.js'
-export * from './dwn-proxy-cli.js'
 export * from './dwn-proxy.js'
 export * from './dwn-types.js'
